Use a list of cached open owners and maintain a refcount
The old implementation of cached open owners allowed refcount to go
to zero, but this causes issues. Instead, if an open owner is to
be cached, put it on a list. If it is re-used before it expires,
pull it off the list. The reaper thread then just walks the list.

Change-Id: Ie72e752e19627b383a4cd1ba47f1c340586a2f69
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
(cherry picked from commit 50d69dd11e3f5210e21715b429d1df0ff3f36e1c)