RPCSEC_GSS: When using kerberos validate principals but not handles
The original issue and the fix were reported in the below link -
        - https://github.com/nfs-ganesha/nfs-ganesha/issues/59

As mentioned there, for auth_gss, its not correct to do memcmp
for comparing credentials since gss_context_id, which is a handle,
shall change.

From RFC 3530, page 67:
>>>>>
As a security measure, the server MUST NOT cancel a client's leased
   state if the principal established the state for a given id string is
   not the same as the principal issuing the SETCLIENTID.
......
......
Note that if the id string in a SETCLIENTID request is properly
constructed, and if the client takes care to use the same principal
for each successive use of SETCLIENTID, then, barring an active
denial of service attack, NFS4ERR_CLID_INUSE should never be
returned.

However, client bugs, server bugs, or perhaps a deliberate change of
the principal owner of the id string (such as the case of a client
that changes security flavors, and under the new flavor, there is no
mapping to the previous owner) will in rare cases result in
NFS4ERR_CLID_INUSE.
<<<<

Hence the fix is to compare kerberos princials instead.

Author:    Alexander Bersenev <bay@hackerdom.ru>
Tested-by: Riyas Abdulrasak <rnalakka@redhat.com>
BUG: 1269169
Change-Id: Ib1b7fe9daf5917636b7ea891a24be2cd0cfc0327
Signed-off-by: Soumya Koduri <skoduri@redhat.com>
(cherry picked from commit 3a34bedef9dc3796ee7a55e3ba7d1966f2adb2a9)