Fix checking for existence of prior locks.
Check for existence of prior locks for pinref before merging locks.
For example, if there is already a lock in write mode and another
locks request in read mode from same owner will make the lock list empty
after the merging locks, but this should not place an additional pinref
on the cache inode entry. Re-arranged the code to have desired effect!

Change-Id: Id7f8cd6a46e2ebdee46cab7a85cab47abbf74933
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>