Fix nlm client refcount going up from zero
Once a refcount goes to zero, we should not be using it any more and let
the thread that made it zero clean up the entry. If zero refcount entry
is found in the hash table, don't use it, wait for the other thread to
clean up and then create new one once the old entry is removed from
the hash table.

Change-Id: Ia10edf32fa40ce31fd53963cec11e6f95b69c0d0
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>