Fix READDIR cookie traversal during deletes
Changes backported from upstream commit
40f8d9c28379b1831de7a8b094eaa09a9eda7a79
If a READDIR is in process while deletes are happening on the same
directory, the return could be BAD_COOKIE when it should have been EOD.
This can happen in the case where all the files after the cookie were
deleted.

Another case is that if the file was deleted, but subsequent files
existed, we might have returned EOD instead of the next dirent.

Change-Id: I9460ad24e154bfb8ba892a8610b3a07fc03ab5e2
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>