Release file system after it is unclaimed
Currently in GPFS after last export of a file system
is removed, the file system is unclaimed but not removed.
If a file system is deleted and a new file system is
added which has the same dev major, minor numbers as the
previous file system then it is skipped as duplicate file
system. Made changes by calling "release_posix_file_system"
after the file system is unclaimed.

Change-Id: I71c8a7e73398ab79ad97e7a11ce09baee676519c
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>