Fix infinite loop in cache_inode_unexport for STALE entry
Currently cache_inode_unexport() gets reference by calling
cache_inode_lru_ref() with flag as LRU_FLAG_NONE. If a
cache entry is in LRU_ENTRY_CLEANUP the cache_inode_lru_ref() returns
error CACHE_INODE_ESTALE. On receiving this error the
cache_inode_unexport() is not able to work on the entry and it
continues working on the same entry in a while() loop resulting in an
infinite loop.
Fixed this by changing cache_inode_unexport() to call
cache_inode_lru_ref() with flag LRU_REQ_STALE_OK. With this flag
being passed to cache_inode_lru_ref(), the function returns a reference
to the entry even if it is in LRU_ENTRY_CLEANUP. With this change, the
cache_inode_unexport() always gets a reference and is able to work with it.

Change-Id: I488bf2082cb4917dbfd1981b9c6c142aacab254b
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>