Determine size of ACLs to be encoded when getting ACLs
Ig nfs4_getfacl is run by the client when
there are large number of ACLs and if the server is not
able to fit in all ACLs in the buffer of size
NFS4_ATTRVALS_BUFFLEN then the client gets Input/Output
error.
To avoid this we now calculate at run time the size
required by ACLs, to determine the size of buffer where
the ACLs would be encoded.

Change-Id: I8b3265237f56c3d9322f4bb641747b305719e4d8
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>