Fix re-use of wrong rpc_dplx_rec leading to a crash.
We create fd and then create rpc_dplx_rec. While releasing we close fd
and then release rpc_dplx_rec. This may make rpc_dplx_rec be re-used for
wrong fd leading to a TCP connection using a UDP rpc_dplx_rec and
ultimately resulting in a crash.

A generation number is atomically incremented when an "fd" is produced.
The hash key used is the generation number and fd. This should prevent
incorrect reuse of rpc_dplx_rec. Changes are extensive but mostly
trivial.

Change-Id: I6c7c65fa0a695b2d02495fe023f1a94a156aa906
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>