[GPFS] Fix upcall to update group, mode when flag UP_OWN is seen
Currently GPFS upcall thread updates the user id but not the group id
when flag UP_OWN is received. With this when change to group for a file
is made on the server side, the client is not able to see the updated
group immediately.
Fix for this is to update the group, mode in the GPFS upcall code when
UP_OWN flag is received.

Change-Id: Iaac4ef05e629991620ecee0fbdccc85402b3a75b
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>