Set attributes_follow=false when required attributes are not filled
Currently in nfs3_FSALattr_To_Fattr() when required attributes are
not filled we log a CRIT message, but do not tell the caller of
nfs3_FSALattr_To_Fattr() about the error. With this the caller sets
attributes_follow=true in the post_op_attr structure even in case of
partial or no attributes. With this there are chances that client will
see attributes_follow=true with partial or no attributes.
To avoid this we now return 'false' from nfs3_FSALattr_To_Fattr() to
tell that required attributes are not filled, thus letting the caller
take appropriate action for the same.

Change-Id: Ic2a1a994165664945161e88201b10f623b46b5b9
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
(cherry picked from commit 1703bd012345ad6460f352c630f294a2d5ef0081)