Move open_fd_count from MDCACHE to FSALs.
FSAL needs to close a file descriptor, if needed, and then open. It is
hard to pass this information back to the upper layers. It is lot easier
to modify this counter in the FSAL itself than in other layers. Of
course, MDCACHE needs to use it to close files though.

Change-Id: I1eea7c57d8dba88caf4f632958f714c0e7403b1c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>