[cmake] Fail cmake if MSPAC feature can't be included
We require MSPAC feature, so fail cmake if we can't include it. A better
approach would be to fail cmake if any feature requested at the "cmake"
CLI invocation can't be included due to needed packages/libraries
missing on the build machine. This simple approach is enough for us for
now.

Change-Id: Ia7efb8fb8e9f1aa7e818c22bea954f5618e8190d
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>