Handle FSAL cookie irregularities in MDCACHE
While reading a directory, if the directory contents are getting
modified simultaneously, then there are chances of hitting cookie
irregularities. Function mdcache_avl_qp_insert returns error code
-3 & -4 in such cases. Error code -3 is returned when file name is
matching with existing name but the cookie is different. This can
happen if new entries got added to the directory resulting in new
cookie for the same file. In case of -3 error code, the client
should reread the directory. Error code -4 is handled by commit:
8f9c8c1a2fa9a84b7fd692d8b6f5451b626bbb91

Change-Id: I2b7fcd20d49f022cc8ec0dd1cd4fde9aab626d6f
Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>