Send correct error code in NFS GETATTR response.
cache_entry_to_nfs3_Fattr() returns a boolean. This is used to send or
not send attributes when attributes are optional. We need to send an
error in response to nfs3_getattr request if we can't provide
attributes! Currenly the code just sends NFS3ERR_INVAL as
cache_entry_to_nfs3_Fattr() just return False when things fail.

Let nfs3_getattr() call the actual underlying procedures that retrieve
the correct status code to send in response.

Change-Id: I3ba01c928b9b10c45de33d69cf19851cb02d3239
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>