Do not store transient open flags with fd
While opening a file, openflags are being passed to FSAL layer and those
flags are stored with fd. The stored flags are used in the reopen call.
As the transient flags are for initial open, those should not be stored.
e.g. FSAL_O_TRUNC flag for initial open should not be stored, else
reopen truncates the file.

Change-Id: I88200888bf59f73829406a17ff660fd04875fda2
Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>