Unset O_TRUNC flag before reopen
The openflags are stored with fd, and those are used while making reopen
call. If the initial open was with O_TRUNC option, then the next reopen
actually truncates the file due to use of old flags. Ideally the next
reopen should not consider the transient flag like O_TRUNC.

Signed-off-by: Sachin Punadikar <psachin@in.ibm.com>