Skip sending the conflicting lock owner in the NLM TEST response
There is a bug found in Linux kernel NFS client code while decoding
the response received from the Ganesha server. As a result locktest
suite is failing. This code change will help use NLM TEST till the
bug is ideally fixed in the relevant XDR routines and released in
kernel NFS clients. Assumption is, it is more important to know
that TEST requests fail rightfully and not who exactly is the
conflicting owner.

Change-Id: I8c00d90b689782d07d94894cd84481a677620415
Signed-off-by: Trishali Nayar <ntrishal@in.ibm.com>