Release dupreq's ref on drc just before freeing the dupreq.
Currently it is released when the dupreq is taken out of the hash table,
but we may access drc from dupreq when it is not in the hash table.

A request gets a ref on drc as well as dupreq. At the end of processing
the request, we release the ref on dupreq and let the dupreq release the
ref on drc as dupreq has access to drc. So it makes sense to release the
ref on drc when we are about to free the dupreq.

Change-Id: I205201b7b8b8bd9d862545d0915178d5015b21ea
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>