Temporary Fix: Don't use maxfilesize for lock byte-range comparison
Currently for LOCKT, LOCK and LOCKU requests we are checking
if requested lock's "offset+length > maxfilesize". If yes, then
request is failed with error NFS4ERR_BAD_RANGE.
For GPFS the "maxfilesize=INT64_MAX". But GPFS allows lock byte-range
upto UINT64_MAX. So we should be comparing requested lock's
"offset+length" with UINT64_MAX and not maxfilesize (INT64_MAX).
This is a temporary fix to change nfs4_op_lock(), nfs4_op_lockt() and
nfs4_op_locku() to check if requested lock's "offset+length > UINT64_MAX"
Once we have a proper fix for this issue this patch will be reverted.

Change-Id: I94d45cc0290dca5700c5a87f42c5b442ebf161f4
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>