Don't increase errcnt when getting obj/export fails in release_openstate()
Currently in release_openstate() in while(errcnt <STATE_ERR_MAX) loop
when 'state' is being removed for an open owner, we try to get reference
to the 'obj' and 'export' associated with the 'state' by calling
get_state_obj_export_owner_refs(). But if get_state_obj_export_owner_refs()
fails and returns with obj/export as NULL we assume that the file, state or
export is about to die and increase the 'errcnt' and continue in the
while(errcnt <STATE_ERR_MAX) loop. Once the 'errcnt' reaches STATE_ERR_MAX
ganesha exits with a FATAL trace. But it is possible that another thread who
is using the same 'obj' associated with the 'state' may be doing time
consuming operation like commit and thus the 'obj' cannot be closed.
In that case while in one thread the 'obj' related to the 'state' is being
used and cannot be removed/closed, in the other thread where release_openstate()
is executing the 'errcnt' quickly increases while running in the while() loop.
So the problem here is ganesha moves very fast through the while() loop in
release_openstate() and if 'obj' doesn't get closed in STATE_ERR_MAX iterations
it exits with FATAL trace. We are not giving enough time for the operation
associated with the 'obj' related to the 'state' to be completed in the other
thread.
To fix this we don't increase the 'errcnt' in release_openstate() when
get_state_obj_export_owner_refs() fails, with this change we are making the code
similar to 2.5 code where 'errcnt' was not increased when
get_state_obj_export_owner_refs() failed.

Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>