Don't unlock 'content_lock' prior to using dirent
In mdcache_readdir_chunked() code, if getattrs() fails for an entry
we unlock the 'directory->content_lock' and then access the 'dirent'.
It is possible that after the 'content_lock' is released another thread
may acquire the same lock  and release the 'dirent'. This can cause a
crash as 'dirent' would be used in mdcache_readdir_chunked() after it
is freed.
To fix this we now unlock the 'content_lock' after we are done using
the 'dirent' and its associated 'entry' in mdcache_readdir_chunked().

Change-Id: I2a25efe9b77c59b64b9946e317ddc6adc76ef1e7
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>