When state lock is held, set no_cleanup=true prior to calling fsal_open2(..)
Currently in open4_ex(..), prior to calling fsal_open2(..) we may have
state_lock_held=true. fsal_open2(..) leads to further call to a FSALs
open2(..), but if it it fails with ESTALE then
mdcache_kill_entry(..) would get called. This may lead to a crash if
there is a call to state_wipe_file(..) where we would try to take the
same state_lock again.
To avoid this we now set state_hdl->no_cleanup=true in open4_ex(..) if
state_lock_held=true.

Change-Id: Icfa2d6abfcf4fd822ab0029c0a3e00c229a96a22
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>