Set no_cleanup=true prior to calling fsal_open2()/fsal_reopen2()
After a "state_lock" is held, if a FSALs open2(..)/reopen2(..) gets
called and if it fails with ESTALE then there is a possibility of
having a crash as we would try to take the same "state_lock" again in
state_wipe_file(..).
To avoid this we now set state_hdl->no_cleanup=true after holding a
"state_lock".

Change-Id: I99c611c14123266a5056c24c9a26e724476d09dd
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>