[GPFS] Don't do permission check after file has been created
In gpfs_open2(), we create a file with the caller's
credentials active, so as such permission check is done.
So we don't need the caller of gpfs_open2() to do permission
check again. For this we avoid passing caller_check_perm to
open_by_handle() so that its value (false) is not modified.

Change-Id: I74031d74fd7c027654c54c9c73a7318818037a9b
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
(cherry picked from commit a20fe0673384023e690582bcbbcd1c3f5738f7ee)