Remove python-qt from dependency for debian builds
We use USE_GUI_ADMIN_TOOLS=NO and don't use any python GUI tools, so
remove the explicit python-qt4 dependency. python-qt4 is not present in
Ubuntu 20.04, hence this change also helps to install ubuntu 18.04 rpms
on ubuntu 20.04 release, obviating the need to build two different sets
of packages (one for 18.04 and another for 20.04).

Change-Id: I5c128fd479eb389b4d04cd13853a9cdcb50893af
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>