Fix stall when readdir races with I/O and a new mdcache entry created
If readdir causes a new mdcache entry to be created and that races
with I/O on another instance of the cache entry, which thus requires
merge_share to resolve the share reservation, the need to take the
obj_lock for write blocks while the I/O completes while holding the
lock for read.

In this case, there's actually no need to do anything in merge_share
because all the counters in the readdir's copy of the obj handle are
0. Since that copy of the obj handle MUST not accessible to any other
thread, it can be accessed without holding the lock.

So we move acquiring and releasing the lock into merge_share and
check that the dupe_share has non-zero counters before taking the
lock.

Change-Id: I372713360a5c795f77b1365c263c66302ec2490a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
(cherry picked from commit aadb9cf27d8846a51cddd4d88f8ebb3ae20a5248)

Conflicts:
	src/FSAL/FSAL_KVSFS/kvsfs_handle.c
	src/FSAL/FSAL_LIZARDFS/handle.c

(cherry picked from commit 85caedb446d4a3762dcbe769e75490d4c06696b8)